- working on getting terrain in

+++ include/terrain/IFieldFunction.hpp
@@ -17,7 +17,7 @@ public:
 	}
 	;
 
-	virtual glm::detail::float32 getNoise(glm::detail::int32 x, glm::detail::int32 y, glm::detail::int32 z) = 0;
+	virtual glm::detail::float32 getNoise(glm::detail::float32 x, glm::detail::float32 y, glm::detail::float32 z) = 0;
 };
 
 }
